+++ TeePee/Examples/TeePee.Examples.WebApp.Tests/FixtureScopedTests/HttpClientFactoryNamedUsageControllerTests.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System.Net;
 using TeePee.Examples.WebApp.Controllers;
+using TeePee.UsageExtensions;
 
 namespace TeePee.Examples.WebApp.Tests.FixtureScopedTests
 {
