+++ TeePee/Examples/TeePee.Examples.WebApp.Tests/TestScopedTests/HttpClientFactoryBasicUsageControllerTests.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 using System.Net;
 //using TeePee.DependencyInjection;
 using TeePee.Examples.WebApp.Controllers;
+using TeePee.UsageExtensions;
 
 namespace TeePee.Examples.WebApp.Tests.TestScopedTests
 {
