+++ TeePee/TeePee.Tests/RuleUsageTests.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using TeePee.Tests.TestData;
+using TeePee.UsageExtensions;
 
 namespace TeePee.Tests;
