+++ TeePee/TeePee/PrimitiveExtensions/StringExtensions.cs
