+++ TeePee/TeePee/RequestMatchBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text;
 using System.Text.Json;
+using TeePee.Built;
 using TeePee.Extensions;
-using TeePee.Internal;
 
 namespace TeePee
 {
